Creating a todo list with MUI
