+++ package.json
@@ -12,6 +12,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "typeface-roboto": "^1.1.13",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
