+++ src/App.css
@@ -1,4 +1,4 @@
-.App {
+/* .App {
   text-align: center;
 }
 
@@ -13,7 +13,7 @@
   }
 }
  input {
-   color: white !important;
+   color: black !important;
  }
 
 .App-header {
@@ -38,4 +38,4 @@
   to {
     transform: rotate(360deg);
   }
-}
+} */
