+++ src/components/Signin.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Signin = () => {
+  console.log("Hello");
+  return (
+    <div>
+      Signin In
+    </div>
+  )
+}
+
+export default Signin;
